+++ packages/client/src/utils.js
@@ -43,8 +43,8 @@ export const isAddr = (addr) => {
   return noPrefix.length === 16;
 };
 export const formatAddress = (addr) => {
-  return addr.startsWith("0x")? addr: `0x${addr}`;
-}
+  return addr.startsWith("0x") ? addr : `0x${addr}`;
+};
 export const formatActionString = (str) => {
   const isFungibleTransfer = str.includes("FungibleToken.Receiver");
   let newStr = "";
